requirements.txt
pandas
pdfplumber
google-generativeai
reportlab
